docs: snippets
