+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "aiola-python"
+name = "aiola"
 version = "0.0.0.dev"
 description = "aiOla python sdk"
 readme = "README.md"
@@ -13,7 +13,7 @@ dev = [
     "pytest-asyncio>=1.0.0",
     "pytest-mock>=3.14.1",
     "ruff>=0.4",
-    # "pyaudio>=0.2.14",
+    # "pyaudio>=0.2.14", # for audio processing, install when using mic-stream
 ]
 
 [tool.ruff]
